requirements.txt
# Build dependencies
setuptools==63.2.0
wheel==0.37.1

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy>=1.26.0,<2.0.0
opencv-python-headless==4.8.1.78
redis==4.5.5
fastapi-limiter==0.1.6
gunicorn==21.2.0

# PyTorch and related
--find-links https://download.pytorch.org/whl/torch_stable.html
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.1+cpu
torchvision==0.17.1+cpu
ultralytics==8.1.28

# Image processing
Pillow==10.0.0

# Data validation and settings
pydantic==2.4.2
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Utilities
requests==2.31.0
tqdm==4.66.1
pandas==2.2.0
seaborn==0.13.2
matplotlib==3.8.2
PyYAML==6.0.1
scikit-learn==1.4.0
onnxruntime==1.21.0